+++ src/dynamics/solver/contact/mod.rs
@@ -6,9 +6,7 @@ mod tangent_part;
 pub use normal_part::ContactNormalPart;
 pub use tangent_part::ContactTangentPart;
 
-use core::cmp::Ordering;
-
-use crate::{dynamics::solver::softness_parameters::SoftnessCoefficients, prelude::*};
+use crate::prelude::*;
 use bevy::{
     ecs::entity::{Entity, EntityMapper, MapEntities},
     reflect::Reflect,
@@ -113,151 +111,6 @@ pub struct ContactConstraint {
 }
 
 impl ContactConstraint {
-    /// Generates a new [`ContactConstraint`] for the given bodies based on a [`ContactManifold`].
-    #[allow(clippy::too_many_arguments)]
-    pub fn generate(
-        manifold_id: usize,
-        manifold: &ContactManifold,
-        body1: &RigidBodyQueryReadOnlyItem,
-        body2: &RigidBodyQueryReadOnlyItem,
-        collider_entity1: Entity,
-        collider_entity2: Entity,
-        collider_transform1: Option<ColliderTransform>,
-        collider_transform2: Option<ColliderTransform>,
-        collision_margin: impl Into<CollisionMargin>,
-        speculative_margin: impl Into<SpeculativeMargin>,
-        friction: Friction,
-        restitution: Restitution,
-        softness: SoftnessCoefficients,
-        warm_start: bool,
-        delta_secs: Scalar,
-    ) -> Self {
-        let collision_margin: Scalar = collision_margin.into().0;
-        let speculative_margin: Scalar = speculative_margin.into().0;
-
-        // Compute the relative dominance of the bodies.
-        let relative_dominance = body1.dominance() - body2.dominance();
-
-        // Compute the inverse mass and angular inertia, taking into account the relative dominance.
-        // TODO: How should we properly take locked axes into account for mass?
-        let (inv_mass1, i1, inv_mass2, i2) = match relative_dominance.cmp(&0) {
-            Ordering::Equal => (
-                body1.mass().inverse(),
-                body1.effective_global_angular_inertia(),
-                body2.mass().inverse(),
-                body2.effective_global_angular_inertia(),
-            ),
-            Ordering::Greater => (
-                0.0,
-                ComputedAngularInertia::INFINITY,
-                body2.mass().inverse(),
-                body2.effective_global_angular_inertia(),
-            ),
-            Ordering::Less => (
-                body1.mass().inverse(),
-                body1.effective_global_angular_inertia(),
-                0.0,
-                ComputedAngularInertia::INFINITY,
-            ),
-        };
-
-        let inverse_mass_sum = inv_mass1 + inv_mass2;
-
-        let mut constraint = Self {
-            index1: *body1.solver_index,
-            index2: *body2.solver_index,
-            collider_entity1,
-            collider_entity2,
-            inverse_mass1: inv_mass1,
-            inverse_mass2: inv_mass2,
-            inverse_inertia1: i1.inverse(),
-            inverse_inertia2: i2.inverse(),
-            relative_dominance,
-            friction,
-            restitution,
-            normal: manifold.normal,
-            points: Vec::with_capacity(manifold.points.len()),
-            manifold_index: manifold_id,
-        };
-
-        let tangents =
-            constraint.tangent_directions(body1.linear_velocity.0, body2.linear_velocity.0);
-
-        for mut contact in manifold.points.iter().copied() {
-            // Transform contact points from collider-space to body-space.
-            if let Some(transform) = collider_transform1 {
-                contact.local_point1 =
-                    transform.rotation * contact.local_point1 + transform.translation;
-            }
-            if let Some(transform) = collider_transform2 {
-                contact.local_point2 =
-                    transform.rotation * contact.local_point2 + transform.translation;
-            }
-
-            contact.penetration += collision_margin;
-
-            let effective_distance = -contact.penetration;
-
-            let local_anchor1 = contact.local_point1 - body1.center_of_mass.0;
-            let local_anchor2 = contact.local_point2 - body2.center_of_mass.0;
-
-            // Store fixed world-space anchors.
-            // This improves rolling behavior for shapes like balls and capsules.
-            let r1 = *body1.rotation * local_anchor1;
-            let r2 = *body2.rotation * local_anchor2;
-
-            // Relative velocity at the contact point.
-            let relative_velocity = body2.velocity_at_point(r2) - body1.velocity_at_point(r1);
-
-            // Keep the contact if (1) the separation distance is below the required threshold,
-            // or if (2) the bodies are expected to come into contact within the next frame.
-            let normal_speed = relative_velocity.dot(constraint.normal);
-            let keep_contact = effective_distance < speculative_margin || {
-                let delta_distance = normal_speed * delta_secs;
-                effective_distance + delta_distance < speculative_margin
-            };
-
-            if !keep_contact {
-                continue;
-            }
-
-            let point = ContactConstraintPoint {
-                // TODO: Apply warm starting scale here instead of in `warm_start`?
-                normal_part: ContactNormalPart::generate(
-                    inverse_mass_sum,
-                    i1,
-                    i2,
-                    r1,
-                    r2,
-                    constraint.normal,
-                    warm_start.then_some(contact.normal_impulse),
-                    softness,
-                ),
-                // There should only be a friction part if the coefficient of friction is non-negative.
-                tangent_part: (friction.dynamic_coefficient > 0.0).then_some(
-                    ContactTangentPart::generate(
-                        inverse_mass_sum,
-                        i1,
-                        i2,
-                        r1,
-                        r2,
-                        tangents,
-                        warm_start.then_some(contact.tangent_impulse),
-                    ),
-                ),
-                max_normal_impulse: 0.0,
-                anchor1: r1,
-                anchor2: r2,
-                normal_speed,
-                initial_separation: -contact.penetration - (r2 - r1).dot(constraint.normal),
-            };
-
-            constraint.points.push(point);
-        }
-
-        constraint
-    }
-
     /// Warm starts the contact constraint by applying the impulses from the previous frame or substep.
     pub fn warm_start(
         &self,
